fix: remove unnecessary u64 cast in replication_e2e test

+++ crates/core/tests/replication_e2e.rs
@@ -229,7 +229,7 @@ fn count_replica_rows(replica_root: &Path, table: &str, column: &str, col_type: 
             let col_path = path.join(format!("{column}.d"));
             if col_path.exists() {
                 let reader = FixedColumnReader::open(&col_path, col_type).unwrap();
-                total += reader.row_count() as u64;
+                total += reader.row_count();
             }
         }
     }
